requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==1.0.1
Werkzeug==3.1.3
zipp==3.21.0
openai==0.28.0
google-api-python-client==2.108.0
google-auth==2.23.4


google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
